% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGOHeatmap}
\alias{plotGOHeatmap}
\title{Plot GO Heatmap}
\usage{
plotGOHeatmap(
  df,
  color.per.type = NULL,
  row.order = NULL,
  col.order = F,
  legend.position = "right",
  legend.justification = c(1, 1),
  legend.title.size = 12,
  legend.key.width = unit(8, "pt"),
  legend.title = "-log10(p-value)"
)
}
\arguments{
\item{df}{columns are cell types, rows are pathways and values are plotted with colors}
}
\description{
Plot GO Heatmap
}
